+++ OUDS/Core/Themes/OrangeCompact/Sources/Values/SemanticTokens/OrangeCompactTheme+FontCompositeSemanticTokens.swift
@@ -21,7 +21,7 @@ import OUDSTokensSemantic
 // swiftlint:disable line_length
 
 /// Defines basic values common to all themes for `FontCompositeSemanticToken`.
-/// These values can be overriden inside ``OrangeCompactThemeFontSemanticTokensProvider`` subclasses (in extensions or not, in the same module or not) thanks to the `@objc public final` combination.
+/// These values can be overridden inside ``OrangeCompactThemeFontSemanticTokensProvider`` subclasses (in extensions or not, in the same module or not) thanks to the `@objc public final` combination.
 /// The aim of this extensions is to make relationships between all composite semantic tokens for typography / fonts and associated composite raw tokens.
 /// The *tokenator* is not able to provide code for such "composite" objects because the *Figma* tool itself cannot manage that and does not output anything in its JSON to process.
 extension OrangeCompactThemeFontSemanticTokensProvider: FontCompositeSemanticTokens {